ai/inference-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing & ML
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4

# Geospatial Processing
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
fiona==1.9.5
pyproj==3.6.1

# HTTP Clients
requests==2.31.0
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async Utilities
celery==5.3.4
redis==5.0.1

# Optional (for specific ML models)
# torch==2.1.0
# tensorflow==2.14.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1